+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 * Add `ds_transaction_id` and `three_d_secure_version` to 3DS pass thru fields
+* Add `cavv`, `xid`, `ds_transaction_id`, `eci_flag`, and `three_d_secure_version`, to `three_d_secure_info`
 
 ## 3.54.0
 * Add `payment_method_nonce` field to `LocalPaymentCompleted` webhook